requirements.txt
Pillow
pycryptodome
requests
